pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mobile-mcp"
version = "0.1.0"
description = "MCP server for mobile app automation — UI control, Frida hooking, traffic interception"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0",
    "uiautomator2>=3.0",
    "frida-tools>=12.0",
    "Pillow>=10.0",
]

[project.scripts]
mobile-mcp = "server:main"
